feat: move the schema definition from 50 to 53
